+++ pkg/cmd/usage_test.go
@@ -1,0 +1,35 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+package cmd
+
+import (
+	"testing"
+
+	"github.com/nirvana-labs/nirvana-cli/internal/mocktest"
+)
+
+func TestUsageList(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	t.Run("regular flags", func(t *testing.T) {
+		mocktest.TestRunMockTestWithFlags(
+			t,
+			"--api-key", "string",
+			"usage", "list",
+			"--max-items", "10",
+			"--cursor", "cursor",
+			"--limit", "10",
+		)
+	})
+}
+
+func TestUsageGet(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	t.Run("regular flags", func(t *testing.T) {
+		mocktest.TestRunMockTestWithFlags(
+			t,
+			"--api-key", "string",
+			"usage", "get",
+			"--resource-id", "123e4567-e89b-12d3-a456-426614174000",
+		)
+	})
+}